Create order class and test along with money for the price.
I decided that although the description of the problem does not mention prices which include fractions of a penny in practice these always seem to crop up so I made amount a BigDecimal.

In real life I would verify whether the assumption that we would only get orders to the nearest penny was true and then use BigDecimal or not as appropriate.